travel_project/requirements.txt
django-admin startproject myproject
cd myproject
python -m venv venv
venv\Scripts\activate
pip install -r requirements.txt
py manage.py createsuperuser